+++ src/components/Hand/index.tsx
@@ -9,6 +9,7 @@ const Hand = (): JSX.Element => {
   const handleMouseDown = () => null;
   return (
     <Style
+      cardCount={handCards.length}
       className={`hand${shouldShowHand ? '' : ' hidden'}`}
       onClick={(event: React.MouseEvent<HTMLInputElement>): void => event.stopPropagation()}>
       {getCards(handCards, handleMouseDown)}