+++ src/components/Hand/style.tsx
@@ -1,11 +1,16 @@
 import styled from 'styled-components';
 
-const Style = styled.div`
+type HandStyle = {
+  cardCount: number;
+};
+
+const Style = styled.div<HandStyle>`
   background: rgba(0, 0, 0, 0.8);
   bottom: 0;
   display: grid;
   grid-row-gap: 8px;
-  grid-template-columns: repeat(10, 175px [col-start]);
+  grid-template-columns: ${props => `repeat(${props.cardCount}, 175px [col-start]);`}
+  justify-content: center;
   left: 0;
   padding: 15px;
   position: fixed;