+++ docs/changelog.md
@@ -2,6 +2,7 @@
 - update jest config.
 - turn off max-lines rule in eslint config.
 - add limit on number of Cards that can be in Hand.
+- center Cards in Hand.
 
 ## Sprint 3
 Ends August 31, 2020